Rename packages to be under `@crowdstrike/*` scope
