+++ docs-app/package.json
@@ -117,15 +117,15 @@
   },
   "dependencies": {
     "@crowdstrike/ember-oss-docs": "^1.1.3",
-    "@ember-headless-form/changeset": "workspace:*",
-    "@ember-headless-form/yup": "workspace:*",
+    "@crowdstrike/ember-headless-form": "workspace:*",
+    "@crowdstrike/ember-headless-form-changeset": "workspace:*",
+    "@crowdstrike/ember-headless-form-yup": "workspace:*",
     "@ember/test-waiters": "^3.0.2",
     "@embroider/router": "^2.0.0",
     "ember-browser-services": "^4.0.4",
     "ember-cached-decorator-polyfill": "^1.0.1",
     "ember-changeset": "^4.1.2",
     "ember-changeset-validations": "^4.1.1",
-    "ember-headless-form": "workspace:*",
     "ember-modifier": "^4.0.0",
     "ember-resources": "^5.4.0",
     "highlight.js": "^11.6.0",
@@ -135,13 +135,13 @@
     "yup": "^1.0.0"
   },
   "dependenciesMeta": {
-    "ember-headless-form": {
+    "@crowdstrike/ember-headless-form": {
       "injected": true
     },
-    "@ember-headless-form/changeset": {
+    "@crowdstrike/ember-headless-form-changeset": {
       "injected": true
     },
-    "@ember-headless-form/yup": {
+    "@crowdstrike/ember-headless-form-yup": {
       "injected": true
     }
   }
