+++ docs/index.md
@@ -16,11 +16,11 @@ order: 1
 To use `ember-headless-form` in your project, run one of the following commands in your terminal:
 
 ```bash
-pnpm add ember-headless-form
+pnpm add @crowdstrike/ember-headless-form
 # or
-yarn add ember-headless-form
+yarn add @crowdstrike/ember-headless-form
 # or
-npm install ember-headless-form
+npm install @crowdstrike/ember-headless-form
 # or
-ember install ember-headless-form
+ember install @crowdstrike/ember-headless-form
 ```
