+++ docs/validation/yup.md
@@ -5,20 +5,20 @@ order: 5
 
 # Integrating with `yup`
 
-[`yup`](https://github.com/jquense/yup) is a popular framework-agnostic validation library. The additional `@ember-headless-form/yup` package makes it easy to integrate `yup` based schema-validation with ember-headless-form. Especially its powerful TypeScript support combined with the [TypeScript / Glint support](../typescript/index.md) of ember-headless-form makes this a compelling solution for TypeScript users, but not only those.
+[`yup`](https://github.com/jquense/yup) is a popular framework-agnostic validation library. The additional `@crowdstrike/ember-headless-form-yup` package makes it easy to integrate `yup` based schema-validation with ember-headless-form. Especially its powerful TypeScript support combined with the [TypeScript / Glint support](../typescript/index.md) of ember-headless-form makes this a compelling solution for TypeScript users, but not only those.
 
 First install both packages:
 
 ```bash
-pnpm add yup @ember-headless-form/yup
+pnpm add yup @crowdstrike/ember-headless-form-yup
 # or
-yarn add yup @ember-headless-form/yup
+yarn add yup @crowdstrike/ember-headless-form-yup
 # or
-npm install yup @ember-headless-form/yup
+npm install yup @crowdstrike/ember-headless-form-yup
 ```
 
 Then you need to define the `yup` schema you want to validate your form data with. Refer to their [documentation](https://github.com/jquense/yup).
 
-Finally we need to integrate the schema with the validation capabilities of the form. The `@ember-headless-form/yup` addon provides a single `validate-yup` helper, that provides the glue code to map `yup` validation errors to the format ember-headless-form expects. Apply it to the [`@validate` form-level callback](./custom-validation.md#form-level-validation), passing the `yup` schema as an argument.
+Finally we need to integrate the schema with the validation capabilities of the form. The `@crowdstrike/ember-headless-form-yup` addon provides a single `validate-yup` helper, that provides the glue code to map `yup` validation errors to the format ember-headless-form expects. Apply it to the [`@validate` form-level callback](./custom-validation.md#form-level-validation), passing the `yup` schema as an argument.
 
 See the following example, where the form's validation is entirely delegated to the `yup` schema:
