+++ packages/changeset/src/helpers/validate-changeset.ts
@@ -1,6 +1,9 @@
 import { isChangeset } from 'validated-changeset';
 
-import type { ErrorRecord, FormValidateCallback } from 'ember-headless-form';
+import type {
+  ErrorRecord,
+  FormValidateCallback,
+} from '@crowdstrike/ember-headless-form';
 import type { EmberChangeset } from 'ember-changeset';
 import { assert } from '@ember/debug';
