+++ test-app/tests/integration/components/yup-test.gts
@@ -4,8 +4,8 @@
 import { click, fillIn, render } from '@ember/test-helpers';
 import { module, test } from 'qunit';
 
-import { HeadlessForm } from 'ember-headless-form';
-import { validateYup } from '@ember-headless-form/yup';
+import { HeadlessForm } from '@crowdstrike/ember-headless-form';
+import { validateYup } from '@crowdstrike/ember-headless-form-yup';
 import sinon from 'sinon';
 import { setupRenderingTest } from 'test-app/tests/helpers';
